infoFeeding/requirements.txt
Django==3.2.9
WeasyPrint==51
gunicorn==20.0.4
djangorestframework==3.12.4
django-cors-headers
